implementar sistema de sesiones

+++ lib/Backend/Feature/Employee/sqlEmployeeDataSource.dart
@@ -6,24 +6,26 @@ import 'package:crud_factories/Objects/Empleoye.dart';
 class SqlEmployeeDataSource implements  IEmployeeDataSource {
 
   final Iexecutequery executeQuery;
+  final String userId;
 
   SqlEmployeeDataSource({
-      required this.executeQuery
+      required this.executeQuery,
+      required this.userId
   });
 
   @override
   Future<void> delete(String id) async {
 
     await executeQuery.execute(
-      'DELETE FROM employees WHERE id=?',
-      [id],
+      'DELETE FROM employees WHERE id= ?  AND user_id = ?',
+      [id,userId],
     );
   }
   @override
   Future<List<Empleoyee>> load() async {
 
     final result = await executeQuery.query(
-      'SELECT id, name,idFactory FROM employees',
+      'SELECT id, name,idFactory FROM employees WHERE userId = ?',
     );
 
     return result.map((row) => Empleoyee(
@@ -38,11 +40,28 @@ class SqlEmployeeDataSource implements  IEmployeeDataSource {
   Future<void> insert(Empleoyee e) async {
 
     await executeQuery.execute(
-        'INSERT INTO employees VALUES (?, ?)',
-        [e.name,e.idFactory,e.id]
+        'INSERT INTO employees VALUES (?, ?,?)',
+        [e.name,e.idFactory,e.id,userId]
     );
   }
 
+
+
+  @override
+  Future<void> save(List<Empleoyee> employees)  async {
+
+    for (final employee in employees) {
+
+      await executeQuery.execute(
+        'INSERT INTO employees VALUES (?, ?, ?, ?)',
+        [
+          employee.id, employee.name, employee.idFactory, userId
+        ],
+      );
+
+    }
+
+  }
 
   @override
   Future<void> upload(Empleoyee empleoyee) {
@@ -54,24 +73,6 @@ class SqlEmployeeDataSource implements  IEmployeeDataSource {
   Future<void> deleteByFactory(String factoryId) {
     // TODO: implement deleteByFactory
     throw UnimplementedError();
-  }
-
-  @override
-  Future<void> save(List<Empleoyee> employees)  async {
-
-    for (final employee in employees) {
-
-      await executeQuery.execute(
-        'INSERT INTO employees (id, name, idFactory) VALUES (?, ?, ?)',
-        [
-          employee.id,
-          employee.name,
-          employee.idFactory,
-        ],
-      );
-
-    }
-
   }
 
 }
