+++ lib/Backend/Feature/Factory/sqlFactoryDataSource.dart
@@ -6,17 +6,18 @@ import 'package:crud_factories/Objects/Factory.dart';
 class SqlFactoryDataSource implements IFactoryDataSource {
 
   final Iexecutequery executeQuery;
+  final String userId;
 
   SqlFactoryDataSource({
-    required this.executeQuery
+    required this.executeQuery, required this.userId
   });
 
   @override
   Future<void> delete(String id) async {
 
       await executeQuery.execute(
-        'DELETE FROM factories WHERE id=?',
-        [id],
+        'DELETE FROM factories WHERE id= ?  AND user_id = ?',
+        [id,userId],
       );
   }
 
@@ -26,7 +27,8 @@ class SqlFactoryDataSource implements IFactoryDataSource {
       final result = await executeQuery.query(
         'SELECT id, name, highDate, sector, telephone1, telephone2, mail, web, address, '
                  'number, apartment,city, province, postcode '
-                                                    'FROM factories'
+            'FROM factories  WHERE userId = ?',
+          [userId]
       );
 
       return result.map((row) => Factory(
@@ -54,9 +56,9 @@ class SqlFactoryDataSource implements IFactoryDataSource {
   @override
   Future<void> insert(Factory f) async {
 
-    await executeQuery.query('INSERT INTO factories values(?,?,?,?,?,?,?,?,?,?,?,?,?,?)',
+    await executeQuery.query('INSERT INTO factories values(?,?,?,?,?,?,?,?,?,?,?,?,?,?, ?)',
       [f.id.toString(),f.name,f.highDate,f.sector,f.thelephones[0],f.thelephones[1],f.mail,f.web,
-       f.address.street,f.address.number,f.address.apartment,f.address.city,f.address.province,f.address.postcode]
+       f.address.street,f.address.number,f.address.apartment,f.address.city,f.address.province,f.address.postcode,userId]
     );
   }
 
@@ -65,9 +67,10 @@ class SqlFactoryDataSource implements IFactoryDataSource {
   Future<void> upload(Factory f) async{
 
     await executeQuery.query('UPDATE factories SET name=?, highDate=?, sector=?, telephone1=?, telephone2=?, mail=?, web=?, '
-        'address=?, number=?, apartment=?, city=?, postCode=?, province=? WHERE id=?',
+        'address=?, number=?, apartment=?, city=?, postCode=?, province=? WHERE id=?  AND user_id = ?',
         [f.name,f.highDate,f.sector,f.thelephones[0],f.thelephones[1],f.mail,f.web,
-        f.address.street,f.address.number,f.address.apartment, f.address.city,f.address.province,f.address.postcode,f.id.toString()]
+        f.address.street,f.address.number,f.address.apartment, f.address.city,f.address.province,f.address.postcode,
+          f.id.toString(),userId]
     );
   }
 
@@ -77,20 +80,9 @@ class SqlFactoryDataSource implements IFactoryDataSource {
       await executeQuery.query(
           'INSERT INTO factories values(?,?,?,?,?,?,?,?,?,?,?,?,?,?)',
           [
-            f.id.toString(),
-            f.name,
-            f.highDate,
-            f.sector,
-            f.thelephones[0],
-            f.thelephones[1],
-            f.mail,
-            f.web,
-            f.address.street,
-            f.address.number,
-            f.address.apartment,
-            f.address.city,
-            f.address.province,
-            f.address.postcode
+            f.id.toString(), f.name, f.highDate, f.sector, f.thelephones[0], f.thelephones[1], f.mail, f.web,
+            f.address.street, f.address.number, f.address.apartment, f.address.city, f.address.province, f.address.postcode,
+            userId
           ]
 
       );
